Fixed query creation in cases when actually a keyword is supplied as term search.
Before, QueryCreator (to be as versatile as possible) was creating too "wide" queries that were actually skewing results, and some hits were unrightfully been scored better as others. An attempt is made to fix it, and combination of KW and non-KW fields are used only, when input term is actually multi termed.

For more and original discussion see
https://issues.sonatype.org/browse/NEXUS-3881


git-svn-id: https://svn.apache.org/repos/asf/maven/indexer/trunk@1039472 13f79535-47bb-0310-9956-ffa450edef68