NEXUS-4149: fixed.
The culprit was the MergedIndexingContext, that when published, resulted in "transport format" (the GZ file) that contained multiple allGroups and rootGroups fields, the one used in "tree view" components (to list root nodes of the index tree). Nexus handled (1.8 tested) and handles this nicely, but m2e was not. It was simply finding _one_ of the present allGroups field (usually the wrong), and presenting tree using it, while the Lucene index itself was correct, and was usable for searches. This affects only "tree view", not the searches.

git-svn-id: https://svn.apache.org/repos/asf/maven/indexer/trunk@1086870 13f79535-47bb-0310-9956-ffa450edef68