MINDEXER-34: Order of IndexCreator's passed to NexusIndexer.addIndexingContext...() matters and may produce wrong results.
Maven Indexer provides a set of "core" creators, and while there is interdependency between them, it was never handled programatically. As Jesse points out in issue, the container returned list (lookupList()) ordering is not guaranteed at all.

Now, the IndexCreator interface is extended to be able to topologically sort creators before they are added to newly created contexts. A runtime exception IllegalArgumentException is newly introduced on NexusIndexer.addIndexingContext... methods, to prevent creation of IndexingContexts with non-satisfiable list of IndexCreators.

git-svn-id: https://svn.apache.org/repos/asf/maven/indexer/trunk@1158739 13f79535-47bb-0310-9956-ffa450edef68