MINDEXER-38: Undoing change sneaked in with revision 1158917. The original reason for fix was the bug when while performing "flat" search, in case 1st searched context had no results, search stopped instead going over to next context.
In the mentioned revision, there was another fix wrt "uniqueness" of the result set made, that is wrong. The passed in Collection<ArtifactInfo> is actually a TreeSet with proper comparator set in FlatSearchRequest. The fix would override other user set comparators, and would break the indexer contract for various use cases.

To demonstrate this, a dedicated UT DuplicateSearchTest is added too, with comprehensive comments why and what happens.

git-svn-id: https://svn.apache.org/repos/asf/maven/indexer/trunk@1159159 13f79535-47bb-0310-9956-ffa450edef68