Getting rid of heap memory hog when unneeded.
DefaultScannerListener hold a Set<String> with all the existing
UINFOs loaded up in initialisation part of processing when
update was happening. The solely purpose of this set was, that
as last step, create diff with processed UINFOs and delete from index
the files not found while scanning but on index. This step makes sense
for contexts that does NOT receive updates from remote, as they are
actually "hosted" reposes (in MRM lingo). In case of "proxy" reposes,
if Indexer did not find a file locally, but is on index, it is
probably a file present on remote (proxied) repository, who's published
index was already downloaded once.

Also, the method removeDeletedArtifacts() making use of this set
was invoked with same condition as the building of set is now
enclosed.

Meaning, the potentially huge Set<String> will not be built anymore
unless really needed and used.

git-svn-id: https://svn.apache.org/repos/asf/maven/indexer/trunk@1373387 13f79535-47bb-0310-9956-ffa450edef68