MINDEXER-64: Update TrueZip to latest stable release
Used TrueZip was a "rc" release. This change updates
it to latest one. There is no API change, but
slightly more excludes are added to prevent
TrueZip Swing related artifact to come in.

Note: new dependency of TrueZip (bouncycastle) is
"shaved" for JAR signature, but only in the shaded
CLI Jar.

Note: TrueZip was always  and is Java6, while MI is Java5!
But as classpath discovery is used to decide is it used or not
this does not oppose problem. Simply, you should include
it if you want to use it, and you are on Java6. On
Java5 you should NOT include it (this is unchanged
and was true for MI 4.x releases too).

git-svn-id: https://svn.apache.org/repos/asf/maven/indexer/trunk@1410244 13f79535-47bb-0310-9956-ffa450edef68