MINDEXER-71 Do not implicitly close indexing context
Indexing context #rollback, #commit and #optimize operations
were implicitely closing indexing context in case of
IOException and CorruptIndexException. This behaviour is
counter-intuitive. It is also harmful when triggered
from within acquireIndexSearcher-finally-releaseIndexSearcher
block, in which case releaseIndexSearcher fails with NPE,
which in turn masks original exception and makes 
troubleshooting of the problem much harder than it should be.



git-svn-id: https://svn.apache.org/repos/asf/maven/indexer/trunk@1451310 13f79535-47bb-0310-9956-ffa450edef68