Fix double invocation of setIndexDirectoryFile
It was always invoked with parameter `null` in
ctor and then potentially involved yet again
with non-null parameter.

If indexDirectoryFile is `null`, then context.id
is used to create temporary directory for context
files (and makes NIO Files.createTempDirectory
fail if context ID contains a file separator).

This change just makes sure that IF indexDirectoryFile
is given, to NOT create temp file just to not
use it, but use provided indexDirectoryFile instead.