Decouple from Plexus
Plexus is not needed anymore for indexer, while
it retains all the functionalities. WagonHelper is
modified to use Injector instead.

Still, plexus is used in UTs/ITs, hence it is
present but only in test scope, as all the "injected
tests" are plexus based, hence redoing that is
much bigger effort.