ArtifactContext should not log warnings if artifacts are no zip files.
DefaultScanner is processing all files which are in the local repo,
which means that ArtifactContext has to deal with files which are
no zip archives too.

This updates ArtifactContext to not log zip exceptions as warning
if it can't open the zip but should still log all other exceptions
which might occur after successful initial read.