Include missing sha1 digests in the index
When other digest types are present in the scanned directory, such as
sha256 and sha512, the sha1 digest is not added to the maven index.

This happens because M2GavCalculator.pathToGav handles all the different
possible hashes: sha1, sha256, sha512, md5 and strips the extension,
resulting in the artifact being classified as having the '.jar'
extension. Now, in DefaultArtifactContextProducer.java, the files with
extension sha1 or md5 are not indexed so the M2GavCalculator doesn't
deal with them. But the other 2 hash types weren't ignored until now and
the M2GavCalculator took the sha256/sha512 hash files and deduced there
must also be a corresponding jar file.

In DefaultScannerListener.artifactDiscovered, the artifacts that were
already processed are skipped by this check:

if (processedUinfos.contains(uinfo)) {
    return; // skip individual snapshots
}

Given the following structure:
qdox-1.5.jar
qdox-1.5.jar.md5
qdox-1.5.jar.sha1
qdox-1.5.jar.sha256
qdox-1.5.jar.sha512

This means that if we process the file qdox-1.5.jar.sha512 before
qdox-1.5.jar, the latter will be skipped entirely. However, the
Sha1Locator cannot find the file named qdox-1.5.jar.sha512.sha1 so it
misses that the artifact qdox-1.5.jar has the sha1 hash stored in
qdox-1.5.jar.sha1, resulting in a missing sha1 hash for this artifact.

The easiest solution is to skip processing sha256 and sha512 files
entirely, as we do for sha1 and md5 files.

Related to #692